requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
langchain-ollama
langchain-core
openai
chromadb
tree-sitter
tiktoken
python-dotenv
rich
